# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kairos - 勤怠管理システム API
  description: |
    Java-based timesheet/attendance management system built with Spring Boot 3.5.3 and Java 21.
    
    ## Features
    - **Timesheet management**: tracking work time, leave types, and report status
    - **Location tracking**: GPS coordinate recording for attendance/timesheet purposes only
    - **JWT Authentication**: Secure API access using JSON Web Tokens
    - **Clean Architecture**: Domain-Driven Design implementation
    
    ## Authentication
    All API endpoints except `/api/auth/**` require JWT Bearer Token authentication.
    
    ## Error Handling
    API returns standardized error responses with appropriate HTTP status codes and Japanese error messages.
  version: 1.0.0
  contact:
    name: Kairos Development Team
    url: https://github.com/okanikani223/kairos
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /api/work-rules/{id}:
    get:
      tags:
        - Work Rules
      summary: 特定勤務ルール取得
      description: 指定したIDの勤務ルールを取得します
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: 勤務ルールID
      responses:
        '200':
          description: 勤務ルール取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkRuleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Work Rules
      summary: 勤務ルール更新
      description: 既存の勤務ルールを更新します
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: 勤務ルールID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkRuleRequest'
      responses:
        '200':
          description: 勤務ルール更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkRuleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Work Rules
      summary: 勤務ルール削除
      description: 指定したIDの勤務ルールを削除します
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: 勤務ルールID
      responses:
        '204':
          description: 勤務ルール削除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Default Work Rules endpoints
  /api/default-work-rules:
    post:
      tags:
        - Default Work Rules
      summary: デフォルト勤務ルール登録
      description: 新しいデフォルト勤務ルールを登録します
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDefaultWorkRuleRequest'
      responses:
        '201':
          description: デフォルト勤務ルール登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultWorkRuleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Default Work Rules
      summary: デフォルト勤務ルール一覧取得
      description: 認証ユーザーのデフォルト勤務ルール一覧を取得します
      responses:
        '200':
          description: デフォルト勤務ルール一覧取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DefaultWorkRuleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/default-work-rules/{id}:
    get:
      tags:
        - Default Work Rules
      summary: 特定デフォルト勤務ルール取得
      description: 指定したIDのデフォルト勤務ルールを取得します
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: デフォルト勤務ルールID
      responses:
        '200':
          description: デフォルト勤務ルール取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultWorkRuleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Default Work Rules
      summary: デフォルト勤務ルール更新
      description: 既存のデフォルト勤務ルールを更新します
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: デフォルト勤務ルールID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDefaultWorkRuleRequest'
      responses:
        '200':
          description: デフォルト勤務ルール更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultWorkRuleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Report Creation Rules endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token認証
  schemas:
...
    UserDto:
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
          description: ユーザーID
          minLength: 1
          example: "user001"

    # Reports DTOs
...
    RegisterWorkRuleRequest:
      type: object
      required:
        - workPlaceId
        - latitude
        - longitude
        - standardStartTime
        - standardEndTime
        - membershipStartDate
        - membershipEndDate
      properties:
        workPlaceId:
          type: integer
          format: int64
          description: 勤怠先ID
          example: 1
        latitude:
          type: number
          format: double
          description: 勤怠先の緯度（-90.0〜90.0）
          minimum: -90.0
          maximum: 90.0
          example: 35.6762
        longitude:
          type: number
          format: double
          description: 勤怠先の経度（-180.0〜180.0）
          minimum: -180.0
          maximum: 180.0
          example: 139.6503
        standardStartTime:
          type: string
          format: time
          description: 規定勤怠開始時刻
          example: "09:00:00"
        standardEndTime:
          type: string
          format: time
          description: 規定勤怠終了時刻
          example: "18:00:00"
        breakStartTime:
          type: string
          format: time
          description: 規定休憩開始時刻
          nullable: true
          example: "12:00:00"
        breakEndTime:
          type: string
          format: time
          description: 規定休憩終了時刻
          nullable: true
          example: "13:00:00"
        membershipStartDate:
          type: string
          format: date
          description: 所属開始日
          example: "2024-01-01"
        membershipEndDate:
          type: string
          format: date
          description: 所属終了日
          example: "2024-12-31"
    UpdateWorkRuleRequest:
      allOf:
        - $ref: '#/components/schemas/RegisterWorkRuleRequest'
    WorkRuleResponse:
      type: object
      required:
        - id
        - workPlaceId
        - user
        - latitude
        - longitude
        - standardStartTime
        - standardEndTime
        - membershipStartDate
        - membershipEndDate
      properties:
        id:
          type: integer
          format: int64
          description: 勤務ルールID
          example: 1
        workPlaceId:
          type: integer
          format: int64
          description: 勤怠先ID
          example: 1
        user:
          $ref: '#/components/schemas/UserDto'
        latitude:
          type: number
          format: double
          description: 勤怠先の緯度
          example: 35.6762
        longitude:
          type: number
          format: double
          description: 勤怠先の経度
          example: 139.6503
        standardStartTime:
          type: string
          format: time
          description: 規定勤怠開始時刻
          example: "09:00:00"
        standardEndTime:
          type: string
          format: time
          description: 規定勤怠終了時刻
          example: "18:00:00"
        breakStartTime:
          type: string
          format: time
          description: 規定休憩開始時刻
          nullable: true
          example: "12:00:00"
        breakEndTime:
          type: string
          format: time
          description: 規定休憩終了時刻
          nullable: true
          example: "13:00:00"
        membershipStartDate:
          type: string
          format: date
          description: 所属開始日
          example: "2024-01-01"
        membershipEndDate:
          type: string
          format: date
          description: 所属終了日
          example: "2024-12-31"

    # Default Work Rules DTOs
    RegisterDefaultWorkRuleRequest:
      type: object
      required:
        - workPlaceId
        - latitude
        - longitude
        - standardStartTime
        - standardEndTime
      properties:
        workPlaceId:
          type: integer
          format: int64
          description: 勤怠先ID
          example: 1
        latitude:
          type: number
          format: double
          description: 勤怠先の緯度（-90.0〜90.0）
          minimum: -90.0
          maximum: 90.0
          example: 35.6762
        longitude:
          type: number
          format: double
          description: 勤怠先の経度（-180.0〜180.0）
          minimum: -180.0
          maximum: 180.0
          example: 139.6503
        standardStartTime:
          type: string
          format: time
          description: 規定勤怠開始時刻
          example: "09:00:00"
        standardEndTime:
          type: string
          format: time
          description: 規定勤怠終了時刻
          example: "18:00:00"
        breakStartTime:
          type: string
          format: time
          description: 規定休憩開始時刻
          nullable: true
          example: "12:00:00"
        breakEndTime:
          type: string
          format: time
          description: 規定休憩終了時刻
          nullable: true
          example: "13:00:00"
    UpdateDefaultWorkRuleRequest:
      allOf:
        - $ref: '#/components/schemas/RegisterDefaultWorkRuleRequest'
    DefaultWorkRuleResponse:
      type: object
      required:
        - id
        - workPlaceId
        - user
        - latitude
        - longitude
        - standardStartTime
        - standardEndTime
      properties:
        id:
          type: integer
          format: int64
          description: デフォルト勤務ルールID
          example: 1
        workPlaceId:
          type: integer
          format: int64
          description: 勤怠先ID
          example: 1
        user:
          $ref: '#/components/schemas/UserDto'
        latitude:
          type: number
          format: double
          description: 勤怠先の緯度
          example: 35.6762
        longitude:
          type: number
          format: double
          description: 勤怠先の経度
          example: 139.6503
        standardStartTime:
          type: string
          format: time
          description: 規定勤怠開始時刻
          example: "09:00:00"
        standardEndTime:
          type: string
          format: time
          description: 規定勤怠終了時刻
          example: "18:00:00"
        breakStartTime:
          type: string
          format: time
          description: 規定休憩開始時刻
          nullable: true
          example: "12:00:00"
        breakEndTime:
          type: string
          format: time
          description: 規定休憩終了時刻
          nullable: true
          example: "13:00:00"

    # Report Creation Rules DTOs
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - message
        - timestamp
      properties:
        errorCode:
          type: string
          description: エラーコード
          enum:
            - VALIDATION_ERROR
            - REQUEST_VALIDATION_ERROR
            - REQUEST_PARSE_ERROR
            - MISSING_PARAMETER
            - JWT_AUTHENTICATION_ERROR
            - AUTHORIZATION_ERROR
            - ACCESS_DENIED
            - RESOURCE_NOT_FOUND
            - METHOD_NOT_ALLOWED
            - DUPLICATE_RESOURCE
            - DATA_INTEGRITY_VIOLATION
            - BUSINESS_RULE_VIOLATION
            - ILLEGAL_STATE_ERROR
            - INTERNAL_ERROR
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: エラーメッセージ（日本語）
          example: "バリデーションエラーが発生しました"
        timestamp:
          type: string
          format: date-time
          description: エラー発生日時
          example: "2024-01-15T10:30:00"
  responses:
    BadRequest:
      description: Bad Request - バリデーションエラーまたはリクエスト形式エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationError:
              summary: バリデーションエラー
              value:
                errorCode: "VALIDATION_ERROR"
                message: "ユーザーIDは必須です"
                timestamp: "2024-01-15T10:30:00"
            parseError:
              summary: リクエストパースエラー
              value:
                errorCode: "REQUEST_PARSE_ERROR"
                message: "リクエストボディの形式が正しくありません"
                timestamp: "2024-01-15T10:30:00"
    Unauthorized:
      description: Unauthorized - 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errorCode: "JWT_AUTHENTICATION_ERROR"
            message: "JWTトークンが無効です"
            timestamp: "2024-01-15T10:30:00"
    Forbidden:
      description: Forbidden - 権限不足エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errorCode: "AUTHORIZATION_ERROR"
            message: "このリソースにアクセスする権限がありません"
            timestamp: "2024-01-15T10:30:00"
    NotFound:
      description: Not Found - リソース不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errorCode: "RESOURCE_NOT_FOUND"
            message: "指定されたリソースが見つかりません"
            timestamp: "2024-01-15T10:30:00"
...
    UnprocessableEntity:
      description: Unprocessable Entity - 業務ルール違反
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errorCode: "BUSINESS_RULE_VIOLATION"
            message: "業務ルールに違反しています"
            timestamp: "2024-01-15T10:30:00"
    InternalServerError:
      description: Internal Server Error - サーバー内部エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errorCode: "INTERNAL_ERROR"
            message: "予期しないエラーが発生しました"
            timestamp: "2024-01-15T10:30:00"